First commit of the SqlBasedIterator. Removed the inner class from the SqlBasedUriCollector
